requirements.txt
simplejson
paho-mqtt
pytz
aiohttp